Moved code to scripts folder
Marked out where to update code to allow for pseudo random

cleaned up code
