--- Assets/Scenes/TEST - Editor Only/Jyama/Throw/Trajectory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e21d1bea26203ec469cc54fe7afa37c5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
