+++ Assets/Scripts/Throw.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a103388fe3fb1fc488c97ed24f2338b8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
