+++ Assets/Scripts/Throw/Throw.cs
@@ -1,0 +1,186 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.InputSystem;
+public class Throw : Base_InputSystem
+{
+    [SerializeField] private GameObject _indicator;
+    [SerializeField] private GameObject[] _throwPrefabs;
+
+    [Space(5)]
+    [SerializeField] private float _force = 20f;
+    [SerializeField] private Vector3 _offset;
+
+    [Space(5)]
+    [SerializeField] private int _linePoints = 25;
+    [SerializeField] private float _totalTime = 2f;
+    [SerializeField] private LayerMask _layerMask;
+
+    [Space(5)]
+    [SerializeField] private float _clampVertical = 85f;
+    [SerializeField] private float _clampHorizontal = 85f;
+    [SerializeField] private float _rotationSpeed = 5f;
+
+    private LineRenderer _lineRenderer;
+
+    private float mass, drag;
+    private GameObject current;
+
+    private bool isHeld = false;
+    private Vector2 _currentRotation;
+
+    private Vector3 _forceVector;
+    private Vector3 startOffset => CalcOffset(Camera.main.transform, _offset);
+
+    
+
+    private void Start()
+    {
+        _lineRenderer = gameObject.GetComponent<LineRenderer>();
+
+        _input.Player.Fire.started += Fire_started;
+        _input.Player.Fire.canceled += Fire_canceled;
+
+        transform.position = CalcOffset(Camera.main.transform, _offset);
+        transform.SetParent(Camera.main.transform, true);
+    }
+
+
+
+    private void Update()
+    {
+        if (isHeld)
+        {
+            Vector2 mouseDelta = Mouse.current.delta.ReadValue();
+            _currentRotation.y += mouseDelta.x * Time.deltaTime * _rotationSpeed;
+            _currentRotation.x += -mouseDelta.y * Time.deltaTime * _rotationSpeed;
+
+            _currentRotation.x = Mathf.Clamp(_currentRotation.x, -_clampVertical, _clampVertical);
+            _currentRotation.y = Mathf.Clamp(_currentRotation.y, -_clampHorizontal, _clampHorizontal);
+
+            transform.localRotation = Quaternion.Euler(_currentRotation.x, _currentRotation.y, 0);
+
+            _forceVector = _force * transform.forward;
+
+            DrawWithDrag(_forceVector);
+            _indicator.transform.position = _lineRenderer.GetPosition(_lineRenderer.positionCount - 1);
+            current.transform.position = startOffset;
+        }      
+    }
+
+    #region Input events for start / end click
+    private void Fire_started(InputAction.CallbackContext obj)
+    {
+        isHeld = true;
+
+        _lineRenderer.enabled = true;
+        _indicator.SetActive(true);
+
+        transform.rotation = Camera.main.transform.rotation;
+        _currentRotation = Vector2.zero;
+
+        SpawnNextThrownObject();
+        UpdateLineColor(current);
+    }
+
+    private void Fire_canceled(InputAction.CallbackContext obj)
+    {
+        isHeld = false;
+
+        ThrowItem(current);
+
+        _lineRenderer.enabled = false;
+        _indicator.SetActive(false);
+    }
+    #endregion
+
+    private void UpdateLineColor(GameObject obj)
+    {
+        Color color = obj.GetComponent<Renderer>().material.color;
+        _lineRenderer.material.SetColor("_Color", color);
+    }
+
+
+    #region Spawn Thrown Objects
+    private void SpawnNextThrownObject()
+    {
+        GameObject prefab = GetNextPrefab();
+
+        Rigidbody rigidbody = prefab.GetComponent<Rigidbody>();
+        mass = rigidbody.mass;
+        drag = rigidbody.drag;
+
+        current = Instantiate(prefab, startOffset, Quaternion.identity, transform);
+    }
+
+    private GameObject GetNextPrefab()
+    {
+        //TODO: Replace code here for pseudo random spawner if needed.
+
+        int RandInt = Random.Range(0, _throwPrefabs.Length);
+        return _throwPrefabs[RandInt];
+    }
+    #endregion
+
+    #region Throw Item
+    private void ThrowItem(GameObject go)
+    {
+        EnableThrownObject(go);
+        Rigidbody rb = go.GetComponent<Rigidbody>();
+
+        rb.velocity = Vector3.zero;
+        rb.angularVelocity = Vector3.zero;
+
+        rb.AddForce(_forceVector, ForceMode.Impulse);
+    }
+
+    private void EnableThrownObject(GameObject go)
+    {
+        go.transform.SetParent(null, true);
+    }
+    #endregion
+
+    /// <summary>
+    /// Draws by calc trajectory w/drag & stops on 1st collision or time travelled
+    /// </summary>
+    /// <param name="ForceVector"></param>
+    private void DrawWithDrag(Vector3 ForceVector)
+    {
+        _lineRenderer.positionCount = _linePoints;
+        float time = _totalTime / _linePoints;
+
+        Vector3 currentPos = startOffset;
+        Vector3 currentVelocity = ForceVector / mass;
+
+        for(int i = 0; i < _linePoints; i++)
+        {
+            _lineRenderer.SetPosition(i, currentPos);
+
+            currentVelocity += Physics.gravity * time;
+            currentVelocity *= (1 - drag * time);
+            
+            Vector3 delta = currentVelocity * time;
+            if (Physics.Raycast(currentPos, delta.normalized, out RaycastHit hit, delta.magnitude, _layerMask))
+            {
+                _lineRenderer.SetPosition(i, hit.point);
+                _lineRenderer.positionCount = i + 1;
+                return;
+            }
+
+            currentPos += delta;
+        }
+    }
+
+    /// <summary>
+    /// Helper function to calc Vector3 directional offsets
+    /// </summary>
+    /// <param name="transform"></param><param name="offset"></param>
+    /// <returns></returns>
+    private Vector3 CalcOffset(Transform transform, Vector3 offset)
+    {
+        return transform.position +
+            offset.x * transform.forward +
+            offset.y * transform.up +
+            offset.z * transform.right;
+    }
+}
